Refactor UI components for modular architecture and enhance functionality
- Introduced a modular UI structure with dedicated components for Document Converter, Chat Interface, and Query Ranker.
- Updated README to reflect the new modular UI architecture and its components.
- Implemented content formatting utilities for Markdown and LaTeX rendering.
- Enhanced file validation and threading utilities for improved user experience.
- Added comprehensive styles for a cohesive UI design across components.
- Established a test suite for the new UI components to ensure functionality and reliability.
