+++ src/ui/components/__init__.py
@@ -1,0 +1,1 @@
+"""UI Components package - Modular UI components for the Markit application."""
