+++ src/ui/formatters/__init__.py
@@ -1,0 +1,1 @@
+"""Content Formatters package - Content formatting and rendering utilities."""
