+++ src/ui/styles/__init__.py
@@ -1,0 +1,1 @@
+"""UI Styles package - CSS styles and theme definitions."""
