+++ src/ui/utils/__init__.py
@@ -1,0 +1,1 @@
+"""UI Utils package - Utility functions for UI components."""
